Revert "[find] Use cPickle instead of poorly implemented CSV to save find database"
This reverts commit 56535684e21ac7981ac3c4d712da35c84ea63dfe.
cPickle has turned out to be unstable enough to handle the load of find.py reading and writing.

I'd say we should change find.py to write in defined intervals, and only read on startup. The memory saving of writing the file to the disk instead of holding it in the memory is not big enough if we read and it again on every message
TODO: Change the hand-made CSV implementation to use http://docs.python.org/2/library/csv.html

This fixes issue #160