[rss] Now using intervals
The feed checking function is now run using the interval decorator.
Previously it would just run in a loop in the startrss function, and
every time startrss was run again, it would create another loop.

We don't seem to be able to change the interval while it's running, but
the code is a lot simpler this way.