[rss] Changed DB table structure
- Merged "rss" and "recent" tables into one table named "rss_feeds"

- Renamed site_name and site_url to feed_*

- Fixed bug in regex checking from previous commit

Previously, the module added a new row to the database for each new
article found, even though it was only ever checking the most recent for
each feed. Now that it only stores the most recent, it makes sense to
keep it right in the feed table.

This obviously won't be able to read from existing databases, unless we
add a check for the old table structure in the checkdb() function.