[core] Unify the usage of bot.Willie.debug across core and all modules.
Tags like core, [CORE] and so on are useless, and most modules used the module
name as the tag. Now that tag is given a file that exists, it will use the
relative (to willie) path to that file as the tag. Otherwise it will work as
before. This allows for passing __file__ as the tag.