Delete ad-hoc whitespace escaping in movie.py
As far as I can tell the web.get method handles this correctly via
urllib2.quote anyway. The pre-emptive replacement of " " with "+" in
movie.py results in a bogus doubly-escaped URI generated for a query
with a space, e.g.

.movie Groundhog Day

The resulting URI is http://www.imdbapi.com/?t=groundhog%2Bday, which is
bogus and produces a 'Movie not found!' reply.

Deleting that line and letting web.get handle it internally generates
the correct response.