[web, url] Move header-based encoding detection to web.get
url.py did this really good thing where it would look for the encoding in the
HTTP headers and use them to .decode() the string returned from web.get.

web.get() gained (not too long ago) the ability to decode text itself, but it
was doing it to utf8 by default.

So instead of hardcoding utf8 in web.get, I moved that detection functionality
from url to web, so now all modules using web.get will be able to enjoy it.