[bugzilla, github, reddit, youtube] Remove handlers on shutdown
This solves #615. When reloading these modules, a new compiled regex
object was created and inserted into the url handler set. Because it was
a new object, it didn't replace the old one (even though the pattern was
the same). This meant that every time the module was reloaded, another
version of the function was added, meaning the output would be produced
an additional time.