[weather] Ugly workaround for feedparser unicode bug on Python 3
Unicode is hard, even for developers of widely used libraries like feedparser.
In 2012, someone reported to them it's broken on Python 3.

They fixed it quickly, but it's 2014, and there's no release on the horizon.
That means that while the fix is in their VCS, users won't have a fixed version
anytime soon... That calls for an ugly workaround!

By making feedparser believe it doesn't have chardect available, we skip the
code branch that uses str() incorrectly. This might lead to unexpected results
if Yahoo! ever do something stupid with the encoding of their feed.
I trust them to not do that.

There's no version check to apply the workaround only to affected versions
of feedparser because it's fucking 01:13 in the morning and I need to sleep.

https://code.google.com/p/feedparser/issues/detail?id=384